Meilleure détection du nom du PDF

+++ puppeteer_scrapping/prodouane_vendanges.js
@@ -132,13 +132,17 @@ const fs = require('fs');
         return false;
     });
     await page.waitForTimeout(100);
-    type = 'fournisseur';
-    ret = await fs.rename('documents/declaration_Production-'+session_id[2]+'_recapitulatif_par_fournisseur.pdf', 'documents/production-'+process.env.PRODOUANE_ANNEE+'-'+session_id[1]+'.pdf', (err) => {if (err) return 'ERR';});
-    if (ret == 'ERR') {
-        type = 'apporteur';
-        await fs.rename('documents/declaration_Production-'+session_id[2]+'_recapitulatif_par_apporteur.pdf', 'documents/production-'+process.env.PRODOUANE_ANNEE+'-'+session_id[1]+'.pdf', (err) => {if (err) throw err;});
+
+    if (fs.existsSync('documents/declaration_Production-'+session_id[2]+'_recapitulatif_par_fournisseur.pdf')) {
+        type = 'fournisseur';
+        await fs.rename('documents/declaration_Production-'+session_id[2]+'_recapitulatif_par_fournisseur.pdf', 'documents/production-'+process.env.PRODOUANE_ANNEE+'-'+session_id[1]+'.pdf', (err) => {if (err) return 'ERR';});
     }
-    
+
+    if (fs.existsSync('documents/declaration_Production-'+session_id[2]+'_recapitulatif_par_apporteur.pdf')) {
+        type = 'apporteur';
+        await fs.rename('documents/declaration_Production-'+session_id[2]+'_recapitulatif_par_apporteur.pdf', 'documents/production-'+process.env.PRODOUANE_ANNEE+'-'+session_id[1]+'.pdf', (err) => {if (err) return 'ERR';});
+    }
+
     if(process.env.DEBUG){
       console.log("Téléchargement PDF OK");
       console.log('documents/declaration_Production-'+session_id[2]+'_recapitulatif_par_'+type+'.pdf'+" => "+'documents/production-'+process.env.PRODOUANE_ANNEE+'-'+session_id[1]+'.pdf');
